more price fetching updates
