+++ src/@types/AssetExtended.d.ts
@@ -1,0 +1,5 @@
+import { Asset } from '@oceanprotocol/lib'
+
+interface AssetExtended extends Asset {
+  consumeDetails?: ConsumeDetails
+}
