+++ src/@utils/subgraph.ts
@@ -410,158 +410,6 @@ export async function getPreviousOrders(
   }
 }
 
-function transformPriceToBestPrice(
-  frePrice: AssetsFrePriceFixedRateExchange[],
-  poolPrice: AssetsPoolPricePool[],
-  freePrice: AssetFreePriceDispenser[]
-) {
-  if (poolPrice?.length > 0) {
-    const consumeDetails: ConsumeDetails = {
-      type: 'dynamic',
-      address: poolPrice[0]?.id,
-      value: poolPrice[0]?.spotPrice,
-      ocean: poolPrice[0]?.baseTokenLiquidity,
-      oceanSymbol: poolPrice[0]?.baseToken.symbol,
-      datatoken: poolPrice[0]?.datatokenLiquidity,
-      pools: [poolPrice[0]?.id],
-      isConsumable: poolPrice[0]?.spotPrice === '-1' ? 'false' : 'true'
-    }
-    return price
-  } else if (frePrice?.length > 0) {
-    // TODO Hacky hack, temporary™: set isConsumable to true for fre assets.
-    // isConsumable: 'true'
-    const price: BestPrice = {
-      type: 'fixed',
-      value: frePrice[0]?.price,
-      address: frePrice[0]?.id,
-      exchangeId: frePrice[0]?.id,
-      oceanSymbol: frePrice[0]?.baseToken.symbol,
-      ocean: 0,
-      datatoken: 0,
-      pools: [],
-      isConsumable: 'true'
-    }
-    return price
-  } else if (freePrice?.length > 0) {
-    const price: BestPrice = {
-      type: 'free',
-      value: 0,
-      address: freePrice[0]?.token.id,
-      exchangeId: '',
-      ocean: 0,
-      datatoken: 0,
-      pools: [],
-      isConsumable: 'true'
-    }
-    return price
-  } else {
-    const price: BestPrice = {
-      type: '',
-      value: 0,
-      address: '',
-      exchangeId: '',
-      ocean: 0,
-      datatoken: 0,
-      pools: [],
-      isConsumable: 'false'
-    }
-    return price
-  }
-}
-
-async function getAssetsPoolsExchangesAndDatatokenMap(
-  assets: Asset[]
-): Promise<
-  [
-    AssetsPoolPricePool[],
-    AssetsFrePriceFixedRateExchange[],
-    AssetFreePriceDispenser[],
-    DidAndDatatokenMap
-  ]
-> {
-  const didDTMap: DidAndDatatokenMap = {}
-  const chainAssetLists: any = {}
-
-  for (const ddo of assets) {
-    didDTMap[ddo?.services[0].datatokenAddress.toLowerCase()] = ddo.id
-    //  harcoded until we have chainId on assets
-    if (chainAssetLists[ddo.chainId]) {
-      chainAssetLists[ddo.chainId].push(
-        ddo?.services[0].datatokenAddress.toLowerCase()
-      )
-    } else {
-      chainAssetLists[ddo.chainId] = []
-      chainAssetLists[ddo.chainId].push(
-        ddo?.services[0].datatokenAddress.toLowerCase()
-      )
-    }
-  }
-  let poolPriceResponse: AssetsPoolPricePool[] = []
-  let frePriceResponse: AssetsFrePriceFixedRateExchange[] = []
-  let freePriceResponse: AssetFreePriceDispenser[] = []
-
-  for (const chainKey in chainAssetLists) {
-    const freVariables = {
-      datatoken_in: chainAssetLists[chainKey]
-    }
-    const poolVariables = {
-      datatokenAddress_in: chainAssetLists[chainKey]
-    }
-    const freeVariables = {
-      datatoken_in: chainAssetLists[chainKey]
-    }
-
-    const queryContext = getQueryContext(Number(chainKey))
-
-    const chainPoolPriceResponse: OperationResult<AssetsPoolPrice> =
-      await fetchData(PoolQuery, poolVariables, queryContext)
-
-    poolPriceResponse = poolPriceResponse.concat(
-      chainPoolPriceResponse.data.pools
-    )
-    const chainFrePriceResponse: OperationResult<AssetsFrePrice> =
-      await fetchData(FreQuery, freVariables, queryContext)
-
-    frePriceResponse = frePriceResponse.concat(
-      chainFrePriceResponse.data.fixedRateExchanges
-    )
-
-    const chainFreePriceResponse: OperationResult<AssetsFreePrice> =
-      await fetchData(FreeQuery, freeVariables, queryContext)
-
-    freePriceResponse = freePriceResponse.concat(
-      chainFreePriceResponse.data.dispensers
-    )
-  }
-  return [poolPriceResponse, frePriceResponse, freePriceResponse, didDTMap]
-}
-
-export async function getAssetsPriceList(assets: Asset[]): Promise<PriceList> {
-  const priceList: PriceList = {}
-
-  const values: [
-    AssetsPoolPricePool[],
-    AssetsFrePriceFixedRateExchange[],
-    AssetFreePriceDispenser[],
-    DidAndDatatokenMap
-  ] = await getAssetsPoolsExchangesAndDatatokenMap(assets)
-  const poolPriceResponse = values[0]
-  const frePriceResponse = values[1]
-  const freePriceResponse = values[2]
-  const didDTMap: DidAndDatatokenMap = values[3]
-
-  for (const poolPrice of poolPriceResponse) {
-    priceList[didDTMap[poolPrice.datatoken.address]] = poolPrice.spotPrice
-  }
-  for (const frePrice of frePriceResponse) {
-    priceList[didDTMap[frePrice.datatoken?.address]] = frePrice.price
-  }
-  for (const freePrice of freePriceResponse) {
-    priceList[didDTMap[freePrice.token?.address]] = '0'
-  }
-  return priceList
-}
-
 export async function getSpotPrice(asset: Asset): Promise<number> {
   const poolVariables = {
     datatokenAddress: asset?.services[0].datatokenAddress.toLowerCase()
@@ -575,49 +423,6 @@ export async function getSpotPrice(asset: Asset): Promise<number> {
   )
 
   return poolPriceResponse.data.pools[0].spotPrice
-}
-
-export async function getAssetsBestPrices(
-  assets: Asset[]
-): Promise<AssetListPrices[]> {
-  const assetsWithPrice: AssetListPrices[] = []
-
-  const values: [
-    AssetsPoolPricePool[],
-    AssetsFrePriceFixedRateExchange[],
-    AssetFreePriceDispenser[],
-    DidAndDatatokenMap
-  ] = await getAssetsPoolsExchangesAndDatatokenMap(assets)
-
-  const poolPriceResponse = values[0]
-  const frePriceResponse = values[1]
-  const freePriceResponse = values[2]
-  for (const ddo of assets) {
-    const dataToken = ddo.services[0].datatokenAddress.toLowerCase()
-    const poolPrice: AssetsPoolPricePool[] = []
-    const frePrice: AssetsFrePriceFixedRateExchange[] = []
-    const freePrice: AssetFreePriceDispenser[] = []
-    const pool = poolPriceResponse.find(
-      (pool: AssetsPoolPricePool) => pool.datatoken.address === dataToken
-    )
-    pool && poolPrice.push(pool)
-    const fre = frePriceResponse.find(
-      (fre: AssetsFrePriceFixedRateExchange) =>
-        fre.datatoken.address === dataToken
-    )
-    fre && frePrice.push(fre)
-    const free = freePriceResponse.find(
-      (free: AssetFreePriceDispenser) => free.token.address === dataToken
-    )
-    free && freePrice.push(free)
-    const bestPrice = transformPriceToBestPrice(frePrice, poolPrice, freePrice)
-    assetsWithPrice.push({
-      ddo: ddo,
-      price: bestPrice
-    })
-  }
-
-  return assetsWithPrice
 }
 
 export async function getHighestLiquidityDatatokens(
